escape quotes for POT files
String are stored in the POT file as double quoted strings
Double quotes are escaped with a backslash (similar to c, javascript, and ruby)

https://www.gnu.org/savannah-checkouts/gnu/gettext/manual/html_node/PO-Files.html#PO-Files

When the programming languages read the PO file, it will parse them as double quoted strings.
When javascript parses the code, the escaping will be removed.

For localization to work, both of those results need to be the same.

This change removes escaping that will be removed by the javascript parser
and adds escaping that will be needed for POT/PO/JS files