fix: improve memory usage when loading the gem (#107)
This is a replacement for the performance portion of #66 after the
revert in #106.

This commit only loads the parsers and gettext within the rake task.
This avoids loading the expensive gettext library at "runtime" when you
don't need the parsing, and moves it to "build time". Note that this is
also how gettext_i18n_rails does it with their slim and haml parsers.

Using the script in #66, the memory savings at runtime is reduced
from ~2MB to ~0.03MB.

Co-authored-by: Keenan Brock <keenan@thebrocks.net>
Co-authored-by: Joe Rafaniello <jrafanie@gmail.com>